Move basebackup code out of init file
