+++ test/basebackup/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) 2022 Aiven, Helsinki, Finland. https://aiven.io/
