+++ test/basebackup/test_chunks.py
@@ -1,3 +1,4 @@
+# Copyright (c) 2022 Aiven, Helsinki, Finland. https://aiven.io/
 import os
 import time
 from pathlib import Path
